ViewInfoExtension RenderVueModelScript
