+++ Web/A2v10.Web.Base/Extensions/HtmlExtensions.cs
@@ -17,27 +17,7 @@ namespace A2v10.Web.Base
 		public static MvcHtmlString RenderVueModelScript(this HtmlHelper<ViewInfo> htmlHelper, String fileName = null)
 		{
 			var viewInfo = htmlHelper.ViewData.Model;
-			var sb = new StringBuilder();
-
-			if (fileName != null)
-			{
-				var appKey = AppConfig.AppKey();
-				var appPath = AppConfig.AppPath();
-
-				String layoutScriptPath = HostingEnvironment.MapPath($"{appPath}{appKey}/{viewInfo.Path}/{fileName}.js");
-
-				StringBuilder scriptText = new StringBuilder(File.ReadAllText(layoutScriptPath));
-				scriptText.Replace("$(PageId)", viewInfo.PageId);
-				scriptText.Replace("$(DataModel)", viewInfo?.Scripts?.DataScript);
-				scriptText.Replace("$(BaseUrl)", viewInfo.BaseUrl);
-
-				sb.Append($"<script type=\"text/javascript\">{scriptText.ToString()}</script>");
-			}
-			else if (!String.IsNullOrEmpty(viewInfo?.Scripts?.Script))
-			{
-				sb.Append($"{viewInfo.Scripts.Script}");
-			}
-			return new MvcHtmlString(sb.ToString());
+			return new MvcHtmlString(viewInfo.RenderVueModelScript(fileName));
 		}
 
 		public static MvcHtmlString RenderPartialView(this HtmlHelper<ViewInfo> htmlHelper, String viewName, Object model)
